requirements.txt
django>=4.2
djangorestframework
django-cors-headers
mysqlclient
python-decouple
gunicorn
whitenoise
openai